% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/functions.R
\name{extract_id}
\alias{extract_id}
\title{Extract an ID from an API response}
\usage{
extract_id(
  resp,
  candidates = c("id", "projectId", "_id", "experimentId", "gwasExperimentId")
)
}
\arguments{
\item{resp}{API response list}

\item{candidates}{Candidate field names}
}
\value{
Character scalar id
}
\description{
Tries common id field names used by the platform.
}
