% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/api_client.R
\name{varometer_api_key}
\alias{varometer_api_key}
\title{Get API key}
\usage{
varometer_api_key(api_key = NULL)
}
\arguments{
\item{api_key}{Character scalar.}
}
\description{
Uses explicit \code{api_key} if provided, otherwise reads \code{ENIOS_API_KEY}.
}
\keyword{internal}
